Moved the filtering component to Modules directory
