+++ .env.production
@@ -1,5 +1,4 @@
-NEXT_PUBLIC_ENV_NAME=Production
 NEXT_PUBLIC_API_URL=https://api.devlaunchers.org/
-NEXT_PUBLIC_STRAPI_URL=https://api.devlaunchers.com/
+NEXT_PUBLIC_STRAPI_URL=https://api.devlaunchers.org/
 NEXT_PUBLIC_GOOGLE_AUTH_URL=https://api.devlaunchers.org/connect/google
 NEXT_PUBLIC_DISCORD_AUTH_URL=https://discord.com/api/oauth2/authorize?client_id=709889509864636496&redirect_uri=https%3A%2F%2Fapi.devlaunchers.com%2Fusers%2Fauth%2Fdiscord%2Fcallback&response_type=code&scope=identify
