+++ src/components/modules/MainPage/filtering/FilteringComponent/StyledFilteringComponent.ts
