+++ src/components/modules/MainPage/filtering/FilteringComponent/index.ts
