+++ src/components/modules/MainPage/filtering/ProjectsListItem/StyleProjectListItem.ts
