+++ src/components/modules/MainPage/filtering/ProjectsListItem/index.ts
