+++ src/components/modules/MainPage/filtering/useProject.tsx
@@ -1,5 +1,5 @@
 import { useCallback, useEffect, useState } from "react";
-import agent from "../../../utils/agent";
+import agent from "../../../../utils/agent";
 import { Project, ProjectLite, ProjectParams } from "./project";
 
 export default function useProjects() {
